requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
gunicorn==20.0.4
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
django-heroku==0.3.1
whitenoise==5.0.1
psycopg2==2.8.4
dj-database-url==0.5.0 
